Applied CR suggestion with consensus
