+++ .vib/kibana/cypress/cypress/integration/kibana_spec.js
@@ -1,7 +1,6 @@
 /// <reference types="cypress" />
 
-import { random } from './utils';
-import { closeThePopups, skipTheWelcomeScreen } from './prepare-app-state.js';
+import { closeThePopups, skipTheWelcomeScreen, random } from './utils';
 
 before(() => {
   skipTheWelcomeScreen();
